balls

+++ bot.py
@@ -75,10 +75,11 @@ async def on_message(message): #budda asked for it, feel free to remove or comme
     if not dev_role in message.author.roles:
         with open(os.environ.get('DATA_DIR') + 'filters.txt', 'r') as file:
             for line in file.read().splitlines():
-                matched = all(re.search(flt, message.content, re.IGNORECASE) for flt in line.split())
-                if matched:
-                    await message.delete()
-                    print('Deleted their spam! Matcher was: ' + line)
+                if not line.trim().empty():
+                    matched = all(re.search(flt, message.content, re.IGNORECASE) for flt in line.split())
+                    if matched:
+                        await message.delete()
+                        print('Deleted their spam! Matcher was: ' + line)
                 
 
 # Regex thing, Techjar said to remove it so i did. here it is just in case you want to re-enable
